// web/src/api/generated/models/HeartMomentDetail.ts
export interface HeartMomentDetail {
    /**
     * 
     * @type {AttachmentSummary}
     * @memberof HeartMomentDetail
     */
    attachment: AttachmentSummary | null;
    /**
     * 
     * @type {AuthorSummary}
     * @memberof HeartMomentDetail
     */
    author: AuthorSummary;
    /**
     * 
     * @type {string}
     * @memberof HeartMomentDetail
     */
    authorId: string;
    /**
     * 
     * @type {ResourceCapabilities}
     * @memberof HeartMomentDetail
     */
    capabilities: ResourceCapabilities;
    /**
     * 
     * @type {Date}
     * @memberof HeartMomentDetail
     */
    createdAt: Date;
    /**
     * 
     * @type {HeartEmotion}
     * @memberof HeartMomentDetail
     */
    emotion: HeartEmotion;
    /**
     * 
     * @type {Date}
     * @memberof HeartMomentDetail
     */
    happenedOn: Date;
    /**
     * 
     * @type {string}
     * @memberof HeartMomentDetail
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof HeartMomentDetail
     */
    spaceId: string;
    /**
     * 
     * @type {string}
     * @memberof HeartMomentDetail
     */
    text: string;
    /**
     * 
     * @type {Date}
     * @memberof HeartMomentDetail
     */
    updatedAt: Date;
    /**
     * 
     * @type {number}
     * @memberof HeartMomentDetail
     */
    version: number;
    /**
     * 
     * @type {ContentVisibility}
     * @memberof HeartMomentDetail
     */
    visibility: ContentVisibility;
}